api/requirements.txt
flask
flask-cors
python-dotenv
Authlib